+++ src/main/java/kr/co/conceptbe/auth/application/dto/PurposeResponse.java
@@ -1,4 +1,4 @@
-package kr.co.conceptbe.member.application.dto;
+package kr.co.conceptbe.auth.application.dto;
 
 public record PurposeResponse(
     Long id,
